+++ Package.swift
@@ -54,7 +54,6 @@ let package = Package(
             path: "Features/AppleSystemScreenTime/Sources",
             linkerSettings: [
                 .linkedFramework("AppKit", .when(platforms: [.macOS])),
-                .linkedFramework("ApplicationServices", .when(platforms: [.macOS])),
                 .linkedLibrary("sqlite3", .when(platforms: [.macOS])),
             ]
         ),